+++ src/ast.rs
@@ -230,7 +230,7 @@ pub struct SoftwareTaskArgs {
     pub core: u8,
 
     /// The task capacity: the maximum number of pending messages that can be queued
-    pub capacity: u8,
+    pub capacity: u16,
 
     /// The priority of this task
     pub priority: u8,
